+++ tests/document_test.py
@@ -1,14 +1,14 @@
-from typing import Dict, List
+from typing import List
 from uuid import uuid4
 
 import pytest
 from pytest_httpx import HTTPXMock
 
-from zep_python import NotFoundError
+from tests.fixtures import API_BASE_URL, mock_healthcheck, undo_mock_healthcheck
 from zep_python.models import Document, DocumentCollection
 from zep_python.zep_client import ZepClient
 
-api_base_url = "http://localhost/api/v1"
+_ = mock_healthcheck, undo_mock_healthcheck
 
 mock_collection_id = str(uuid4())
 mock_collection = DocumentCollection(
@@ -71,7 +71,7 @@ def validate__collection(collection: DocumentCollection) -> None:
     assert collection.embedding_dimensions == 768
     assert collection.embedding_model_name == "bert-base-uncased"
     assert collection.distance_function == "cosine"
-    assert collection.is_normalized == True
+    assert collection.is_normalized is True
 
 
 def validate__document(document: Document) -> None:
@@ -84,7 +84,7 @@ def validate__document(document: Document) -> None:
 
 def validate__batchdocument(documents: List[Document]) -> None:
     # Predefined collection and document objects
-    mock_collection = DocumentCollection(
+    DocumentCollection(
         name="mock_collection", description="Mock Collection", id=mock_collection_id
     )
 
@@ -96,7 +96,7 @@ mock_document = Document(id=mock_document_id, data={"key": "value"})
 async def test_add_collection(httpx_mock: HTTPXMock):
     httpx_mock.add_response(status_code=200, json="Collection Added")
 
-    async with ZepClient(base_url=api_base_url) as client:
+    async with ZepClient(base_url=API_BASE_URL) as client:
         response = await client.add_collection(mock_collection)
 
         assert response == '"Collection Added"'
@@ -106,7 +106,7 @@ async def test_add_collection(httpx_mock: HTTPXMock):
 async def test_get_collection(httpx_mock: HTTPXMock):
     httpx_mock.add_response(status_code=200, json=mock_collection.dict())
 
-    async with ZepClient(base_url=api_base_url) as client:
+    async with ZepClient(base_url=API_BASE_URL) as client:
         collection = await client.get_collection(mock_collection_id)
 
         validate__collection(collection)
@@ -116,7 +116,7 @@ async def test_get_collection(httpx_mock: HTTPXMock):
 async def test_update_collection(httpx_mock: HTTPXMock):
     httpx_mock.add_response(status_code=200, json="Collection Updated")
 
-    async with ZepClient(base_url=api_base_url) as client:
+    async with ZepClient(base_url=API_BASE_URL) as client:
         response = await client.update_collection(mock_collection)
 
         assert response == '"Collection Updated"'
@@ -126,7 +126,7 @@ async def test_update_collection(httpx_mock: HTTPXMock):
 async def test_delete_collection(httpx_mock: HTTPXMock):
     httpx_mock.add_response(status_code=200, json="Collection Deleted")
 
-    async with ZepClient(base_url=api_base_url) as client:
+    async with ZepClient(base_url=API_BASE_URL) as client:
         status = await client.delete_collection(mock_collection_id)
 
         assert status == '"Collection Deleted"'
@@ -136,7 +136,7 @@ async def test_delete_collection(httpx_mock: HTTPXMock):
 async def test_get_document(httpx_mock: HTTPXMock):
     httpx_mock.add_response(status_code=200, json=mock_document.dict())
 
-    async with ZepClient(base_url=api_base_url) as client:
+    async with ZepClient(base_url=API_BASE_URL) as client:
         document = await client.get_document(mock_collection_id, mock_document_id)
 
         assert document == mock_document
@@ -146,7 +146,7 @@ async def test_get_document(httpx_mock: HTTPXMock):
 async def test_add_document(httpx_mock: HTTPXMock):
     httpx_mock.add_response(status_code=200, json={"ids": [mock_document_id]})
 
-    async with ZepClient(base_url=api_base_url) as client:
+    async with ZepClient(base_url=API_BASE_URL) as client:
         document_ids = await client.add_document(mock_collection_id, [mock_document])
 
         assert document_ids["ids"] == [mock_document_id]
@@ -156,7 +156,7 @@ async def test_add_document(httpx_mock: HTTPXMock):
 async def test_update_document(httpx_mock: HTTPXMock):
     httpx_mock.add_response(status_code=200, json="Document Updated")
 
-    async with ZepClient(base_url=api_base_url) as client:
+    async with ZepClient(base_url=API_BASE_URL) as client:
         status = await client.update_document(mock_collection_id, mock_modifieddocument)
 
         assert status == '"Document Updated"'
@@ -166,7 +166,7 @@ async def test_update_document(httpx_mock: HTTPXMock):
 async def test_delete_document(httpx_mock: HTTPXMock):
     httpx_mock.add_response(status_code=200, json="Document Deleted")
 
-    async with ZepClient(base_url=api_base_url) as client:
+    async with ZepClient(base_url=API_BASE_URL) as client:
         status = await client.delete_document(mock_collection_id, mock_document_id)
 
         assert status == '"Document Deleted"'
@@ -176,7 +176,7 @@ async def test_delete_document(httpx_mock: HTTPXMock):
 async def test_batchupdate_documents(httpx_mock: HTTPXMock):
     httpx_mock.add_response(status_code=200, json="Batch Updated")
 
-    async with ZepClient(base_url=api_base_url) as client:
+    async with ZepClient(base_url=API_BASE_URL) as client:
         status = await client.batchupdate_documents(mock_collection_id, [mock_document])
 
         assert status == '"Batch Updated"'
@@ -186,7 +186,7 @@ async def test_batchupdate_documents(httpx_mock: HTTPXMock):
 async def test_batchdelete_documents(httpx_mock: HTTPXMock):
     httpx_mock.add_response(status_code=200, json="Batch Deleted")
 
-    async with ZepClient(base_url=api_base_url) as client:
+    async with ZepClient(base_url=API_BASE_URL) as client:
         status = await client.batchdelete_documents(
             mock_collection_id, [mock_document_id]
         )
@@ -198,7 +198,7 @@ async def test_batchdelete_documents(httpx_mock: HTTPXMock):
 async def test_batchget_documents_byid(httpx_mock: HTTPXMock):
     httpx_mock.add_response(status_code=200, json=[mock_document.dict()])
 
-    async with ZepClient(base_url=api_base_url) as client:
+    async with ZepClient(base_url=API_BASE_URL) as client:
         documents = await client.batchget_documents_byid(
             mock_collection_id, mock_getbatchrequest_byid
         )
@@ -211,7 +211,7 @@ async def test_batchget_documents_byid(httpx_mock: HTTPXMock):
 async def test_batchget_documents_byuuid(httpx_mock: HTTPXMock):
     httpx_mock.add_response(status_code=200, json=[mock_document.dict()])
 
-    async with ZepClient(base_url=api_base_url) as client:
+    async with ZepClient(base_url=API_BASE_URL) as client:
         documents = await client.batchget_documents_byuuid(
             mock_collection_id, mock_getbatchrequest_byuuid
         )
